pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "API FastAPI com MongoDB para gerenciamento de servidores de jogos"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=1.26.0",
    "motor>=3.6.0",
    "pymongo>=4.10.1",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "colorlog>=6.10.1",
    "uvicorn[standard]>=0.32.1",
    "fastapi-pagination>=0.15.6",
    "fastapi[standard]>=0.121.0",
    "pydantic>=2.12.4",
    "pytest>=9.0.2",
]
